Renames the stake action pack to "Staking Actions", closes #942
