+++ packages/i18n/locales/en/translation.json
@@ -479,10 +479,11 @@
     "setItemDescription": "Set the value of an item.",
     "spendActionDescription_dao": "Spend native or CW20 tokens from the treasury.",
     "spendActionDescription_wallet": "Spend native tokens from your wallet.",
-    "stakeActionDescription": "Manage native token staking.",
+    "stake": "Stake",
     "stakeYourTokensToJoin": "Stake your ${{tokenSymbol}} tokens to become a member of {{daoName}}.",
     "staked": "staked",
     "stakes": "Stakes",
+    "stakingActionsDescription": "Manage native token staking: claim rewards, delegate, redelegate, and undelegate.",
     "stakingAddress": "Staking Contract",
     "subDaoExplanation": "SubDAOs are children DAOs that help manage groups within your DAO.",
     "subDaoNeedsAdding": "{{parent}} has not yet recognized {{child}} as a SubDAO.",
@@ -727,6 +728,7 @@
     "stake": "Stake",
     "staked": "Staked",
     "stakedTo": "Staked to...",
+    "stakingActions": "Staking actions",
     "status": "Status",
     "subDaos": "SubDAOs",
     "supplies": "Supplies",
