+++ packages/stateful/actions/components/index.tsx
@@ -6,7 +6,7 @@ export * from './Instantiate'
 export * from './RemoveCw20'
 export * from './RemoveCw721'
 export * from './Spend'
-export * from './Stake'
+export * from './StakingActions'
 export * from './NativeCoinSelector'
 export * from './ActionCard'
 export * from './ActionsRenderer'
